Take in account marked for removal kernels.
The calculation how many kernels to remove before this commit was not
taking in account kernels which are already marked for removal. Thus
when adding new kernels for removal the logic would end up removing more
than the user requested.

This commit just adds the already marked for removal kernels to the
calculation.

Should fix #15.